DocuQuiz-AI-Windows/requirements.txt
easyocr
pytesseract 
tox
fpdf2
Pillow
opencv-python
google-cloud-vision
google-auth
langchain
openai
python-dotenv
langchain-community
python-dotenv
openai
tiktoken
faiss-cpu
pypdf
langchain-openai
langchain-cli